WIP (Moved Lesson.js to LessonPage.js, and to the correct folder)
