+++ src/components/LessonPage/Content.scss
