+++ src/components/LessonPage/Progress.js
