+++ src/components/LessonPage/Progress.scss
