+++ src/components/LessonPage/ResetButton.scss
