+++ src/pages/LessonPage.scss
@@ -1,4 +1,4 @@
-@import '../../styles/mixins.scss';
+@import '../styles/mixins';
 @import '~bootstrapStyles/mixins/image';
 
 .container {
